+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 1.2
 ---
 
+ * Moved history tracker of out main service into the main controller
  * Support complete internalization
  * Two grid view on main page
  * Compact announce form view in administration panel